backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.75
langchain-text-splitters==0.3.11
langsmith==0.4.25
lxml==6.0.1
MarkupSafe==3.0.2
numpy==2.3.2
ollama==0.5.3
orjson==3.11.3
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
python-docx==1.2.0
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zstandard==0.24.0